Validator refactor (#478)
* refactor validators to multiple files

* complete validator split
