+++ guardrails/validators/bug_free_python.py
@@ -1,0 +1,44 @@
+import ast
+
+from typing import Any, Dict
+
+from guardrails.validator_base import (
+    FailResult,
+    PassResult,
+    ValidationResult,
+    Validator,
+    register_validator,
+)
+
+from guardrails.logger import logger
+
+
+@register_validator(name="bug-free-python", data_type="string")
+class BugFreePython(Validator):
+    """Validates that there are no Python syntactic bugs in the generated code.
+
+    This validator checks for syntax errors by running `ast.parse(code)`,
+    and will raise an exception if there are any.
+    Only the packages in the `python` environment are available to the code snippet.
+
+    **Key Properties**
+
+    | Property                      | Description                       |
+    | ----------------------------- | --------------------------------- |
+    | Name for `format` attribute   | `bug-free-python`                 |
+    | Supported data types          | `string`                          |
+    | Programmatic fix              | None                              |
+    """
+
+    def validate(self, value: Any, metadata: Dict) -> ValidationResult:
+        logger.debug(f"Validating {value} is not a bug...")
+
+        # The value is a Python code snippet. We need to check for syntax errors.
+        try:
+            ast.parse(value)
+        except SyntaxError as e:
+            return FailResult(
+                error_message=f"Syntax error: {e.msg}",
+            )
+
+        return PassResult()
