+++ guardrails/validators/endpoint_is_reachable.py
@@ -1,0 +1,55 @@
+
+from typing import Any, Dict
+
+from guardrails.validator_base import (
+    FailResult,
+    PassResult,
+    ValidationResult,
+    Validator,
+    register_validator,
+)
+
+from guardrails.logger import logger
+
+
+@register_validator(name="is-reachable", data_type=["string"])
+class EndpointIsReachable(Validator):
+    """Validates that a value is a reachable URL.
+
+    **Key Properties**
+
+    | Property                      | Description                       |
+    | ----------------------------- | --------------------------------- |
+    | Name for `format` attribute   | `is-reachable`                    |
+    | Supported data types          | `string`,                         |
+    | Programmatic fix              | None                              |
+    """
+
+    def validate(self, value: Any, metadata: Dict) -> ValidationResult:
+        logger.debug(f"Validating {value} is a valid URL...")
+
+        import requests
+
+        # Check that the URL exists and can be reached
+        try:
+            response = requests.get(value)
+            if response.status_code != 200:
+                return FailResult(
+                    error_message=f"URL {value} returned "
+                    f"status code {response.status_code}",
+                )
+        except requests.exceptions.ConnectionError:
+            return FailResult(
+                error_message=f"URL {value} could not be reached",
+            )
+        except requests.exceptions.InvalidSchema:
+            return FailResult(
+                error_message=f"URL {value} does not specify "
+                f"a valid connection adapter",
+            )
+        except requests.exceptions.MissingSchema:
+            return FailResult(
+                error_message=f"URL {value} does not contain " f"a http schema",
+            )
+
+        return PassResult()
