+++ guardrails/validators/two_words.py
@@ -1,0 +1,37 @@
+
+from typing import Any, Dict
+
+from guardrails.validator_base import (
+    FailResult,
+    PassResult,
+    ValidationResult,
+    Validator,
+    register_validator,
+)
+
+from guardrails.logger import logger
+
+
+@register_validator(name="two-words", data_type="string")
+class TwoWords(Validator):
+    """Validates that a value is two words.
+
+    **Key Properties**
+
+    | Property                      | Description                       |
+    | ----------------------------- | --------------------------------- |
+    | Name for `format` attribute   | `two-words`                       |
+    | Supported data types          | `string`                          |
+    | Programmatic fix              | Pick the first two words.         |
+    """
+
+    def validate(self, value: Any, metadata: Dict) -> ValidationResult:
+        logger.debug(f"Validating {value} is two words...")
+
+        if len(value.split()) != 2:
+            return FailResult(
+                error_message="must be exactly two words",
+                fix_value=" ".join(value.split()[:2]),
+            )
+
+        return PassResult()
