+++ guardrails/validators/valid_url.py
@@ -1,0 +1,46 @@
+
+from typing import Any, Dict
+
+from guardrails.validator_base import (
+    FailResult,
+    PassResult,
+    ValidationResult,
+    Validator,
+    register_validator,
+)
+
+from guardrails.logger import logger
+
+
+@register_validator(name="valid-url", data_type=["string"])
+class ValidURL(Validator):
+    """Validates that a value is a valid URL.
+
+    **Key Properties**
+
+    | Property                      | Description                       |
+    | ----------------------------- | --------------------------------- |
+    | Name for `format` attribute   | `valid-url`                       |
+    | Supported data types          | `string`                          |
+    | Programmatic fix              | None                              |
+    """
+
+    def validate(self, value: Any, metadata: Dict) -> ValidationResult:
+        logger.debug(f"Validating {value} is a valid URL...")
+
+        from urllib.parse import urlparse
+
+        # Check that the URL is valid
+        try:
+            result = urlparse(value)
+            # Check that the URL has a scheme and network location
+            if not result.scheme or not result.netloc:
+                return FailResult(
+                    error_message=f"URL {value} is not valid.",
+                )
+        except ValueError:
+            return FailResult(
+                error_message=f"URL {value} is not valid.",
+            )
+
+        return PassResult()
